requirements.txt
asgiref==3.3.4
boto3==1.17.55
botocore==1.20.55
Cerberus==1.3.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
jmespath==0.10.0
mysql-connector==2.2.9
mysql-connector-python==8.0.24
protobuf==3.15.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
s3transfer==0.4.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
Werkzeug==1.0.1
stripe==2.60.0
pywebpush==1.13.0
PyMySQL==1.0.2


